data_transfers/requirements.txt
pandas==0.25.1
pyyaml==5.1.2
google-cloud-storage==1.31.0
apache-beam[gcp]==2.30.0